Add tests to CarManager
Set up Mockito-1.10 and JUnit5, created tests methods and implemented some tests for CarManager#remove.

Had to change the class not to use static methods because it was making it difficult to use mocks.

I know you are not using the managers classes but we should, instead of instantiating the Daos inside the view controllers.